backend/requirements.txt
django>=5.1.5
django-cors-headers>=4.3.1
google-generativeai>=0.3.0
